InvalidCardFormatExceptionTest: change expected outcome.
When an empty string is passed in as the detail message,
expect to get the empty String in the detail message
instead of the default detail message.

This is because it might be useful that an attempt was
made to use an empty string to create a card, opposed
to a null string.