GameStateFactoryTest: avoid duplicates in “tooManyPockets” test.
Ensure that the Board in the test fixture does not
contain any cards used to populate the 11 Pockets.

The Pocket cards use indices 0-21, so a collision
is easily avoided by using indices 49-51 for the
3 Board cards.

NOTE: Although cardFromNumber(int) has its own set
of test methods, I’m getting a “smell” by relying
on it in the Test Class setUp() method. My inclination
is that cardFromNumber(int) (and its test methods)
should be moved to a different class. I plan to
address this in a later commit.