Adds test to ensure info of other Player’s is factored correctly.
As compared to the 1 Pocket version of this test,
Player 0’s chances of getting a pair are lowered,
because we now know that some of the other players
are holding “out” cards.

Out of 820 possible combinations of Turn and River
cards, there are 385 that involve pairing one of the
5 seen ranks, and 43 that involve paring on one
of the 8 unseen ranks.

Total of 428 ways to pair out of 820 possible
combinations.