Define HandEvaluator interface.
This is a specific type of Function<T, R> that
accepts collections of Cards and returns a boolean.

ProbabilityCalculator already uses this Function,
but now we have a specific name for it.  This will
make it easier to read containers that hold
HandEvaluators — i.e. a Map<Outcome, HandEvaluator> rather
than a Map<Outcome, Function<Collection<Card>, Boolean>>.