Factor out a generic hasStraight() helper method.
The helper method is primitive. It can only order
cards by Rank in one way — so Aces are either high
or low, but not both.  The Rank ordering is determined
by a Function that is accepted as an argument.

This allows the original hasStraight() method to
simply return true if a straight is found with
either aces high or aces low.

Computationally, this is not as efficient;
organizationally, though, it is easier to follow
and understand.