Re-factor hasStraight() to focus on Rank values.
Focus on Ranks instead of Cards.  This removes a
level indirection (i.e. “rank” instead of
“card.getRank()”) and the groupingBy() method
provides a tidy way to remove duplicates.  Having
a Set of Ranks instead of a Collection lets us
remove the “no-op” else clause that documented
how to handle two cards of the same rank.